requirements.txt
ccxt==4.3.22
cryptofeed==2.4.1
fredapi==0.5.2
imbalanced-learn==0.12.2
mypy==1.10.0
numpy==1.26.4
optuna==3.6.1
pandas==2.3.1
prometheus-client==0.22.1
pydantic==2.11.7
pytest==8.2.2
pytest-asyncio==1.1.0
python-dotenv==1.1.1
pyyaml==6.0.1
requests==2.32.3
httpx==0.28.1
ruff==0.4.5
scikit-learn==1.3.2  # Downgraded for compatibility
prophet==1.1.7
tensorflow==2.16.1

# Enhanced ML Dependencies
lightgbm==4.1.0
xgboost==2.0.3
transformers==4.36.0
torch>=2.0.0,<3.0.0
optimum[onnxruntime]==1.16.0
tokenizers==0.15.0
datasets==2.14.0
accelerate==0.24.0
stable_baselines3==1.8.0

# Financial ML & Time Series
arch==6.2.0  # GARCH models
statsmodels==0.14.0  # HAR-RV and econometric models
scipy==1.11.4
ta-lib==0.4.28  # Technical analysis
yfinance==0.2.18  # Backup data source

# NLP & Sentiment Analysis
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Enhanced Backtesting & Validation
mlfinlab==1.5.0  # Advanced financial ML
pyfolio==0.9.2  # Portfolio analytics
empyrical==0.5.5  # Risk metrics

# Data & Caching
redis==5.0.1  # Caching for data and models
joblib==1.3.2  # Model persistence and parallel processing
dask==2023.12.0  # Large-scale data processing
